Formatting tweaks to order of achievements
Improved achievement listing on user page to not show user image each
line
Tweaked ordering to ensure ordering includes created_at for multiple lan
ordering
User list now shows icons of last 6 achievements

+++ app/Domain/Achievements/AchievementService.php
@@ -115,7 +115,11 @@ class AchievementService extends ResourceService {
 					$newname = $this->generateUniqueName(public_path()."/uploads/achievements/","",Input::file('image_file')->getClientOriginalExtension());
 					if ($path = Input::file('image_file')->move(public_path()."/uploads/achievements/",$newname)) {
 						$model->image = str_replace(public_path(),"",$path);
-						unlink(public_path().$orig);
+						if ($orig != "") {
+							try {
+								unlink(public_path().$orig);
+							} catch (Exception $ex) {}
+						}
 					} else {
 						unlink(public_path()."/uploads/achievements/".$newname);
 					}
